Extract logic to highlight mentions to the mentions module
